Update readme, package name and references
