--- apiary-hive-consumers/pom.xml
@@ -1,25 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <parent>
-        <groupId>com.expediagroup.apiary</groupId>
-        <artifactId>apiary-extensions-parent</artifactId>
-        <version>3.0.1-SNAPSHOT</version>
-    </parent>
-
-    <artifactId>apiary-hive-consumers-parent</artifactId>
-    <packaging>pom</packaging>
-
-    <dependencies>
-        <dependency>
-            <groupId>org.apache.hive</groupId>
-            <artifactId>hive-metastore</artifactId>
-            <version>${hive.version}</version>
-        </dependency>
-    </dependencies>
-    <modules>
-        <module>hive-consumer-common</module>
-        <module>privileges-grantor</module>
-    </modules>
-
-</project>
