--- apiary-hive-consumers/privileges-grantor/pom.xml
@@ -1,31 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <parent>
-        <groupId>com.expediagroup.apiary</groupId>
-        <artifactId>apiary-extensions-parent</artifactId>
-        <version>3.0.1-SNAPSHOT</version>
-    </parent>
-
-    <artifactId>apiary-privileges-grantor-parent</artifactId>
-    <packaging>pom</packaging>
-
-    <modules>
-        <module>privileges-grantor-core</module>
-        <module>privileges-grantor-lambda</module>
-    </modules>
-
-    <properties>
-        <beeju.version>1.3.0</beeju.version>
-    </properties>
-
-    <dependencies>
-        <dependency>
-            <groupId>com.hotels</groupId>
-            <artifactId>beeju</artifactId>
-            <version>${beeju.version}</version>
-            <scope>test</scope>
-        </dependency>
-    </dependencies>
-
-</project>
