--- apiary-hive-consumers/privileges-grantor/privileges-grantor-core/pom.xml
@@ -1,64 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <parent>
-        <groupId>com.expediagroup.apiary</groupId>
-        <artifactId>apiary-privileges-grantor-parent</artifactId>
-        <version>3.0.1-SNAPSHOT</version>
-    </parent>
-
-    <artifactId>apiary-privileges-grantor-core</artifactId>
-
-    <properties>
-        <!--Test -->
-        <mockitoCore.version>2.21.0</mockitoCore.version>
-        <junit.version>4.12</junit.version>
-        <beeju.version>1.3.0</beeju.version>
-
-        <!-- Plugins -->
-        <mavenSurefire.version>2.22.0</mavenSurefire.version>
-
-    </properties>
-
-    <dependencies>
-        <dependency>
-            <groupId>com.expediagroup.apiary</groupId>
-            <artifactId>apiary-hive-consumer-common</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-
-        <!-- Test dependencies -->
-        <dependency>
-            <groupId>org.mockito</groupId>
-            <artifactId>mockito-core</artifactId>
-            <version>${mockitoCore.version}</version>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>junit</groupId>
-            <artifactId>junit</artifactId>
-            <version>${junit.version}</version>
-            <scope>test</scope>
-        </dependency>
-    </dependencies>
-    <build>
-        <plugins>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-surefire-plugin</artifactId>
-                <version>${mavenSurefire.version}</version>
-            </plugin>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-compiler-plugin</artifactId>
-                <configuration>
-                    <source>8</source>
-                    <target>8</target>
-                </configuration>
-            </plugin>
-        </plugins>
-    </build>
-</project>
