+++ apiary-metastore-events/apiary-metastore-consumer/metastore-consumer-common/src/main/java/com/expediagroup/apiary/extensions/events/metastore/consumer/common/exception/HiveClientException.java
@@ -1,0 +1,26 @@
+/**
+ * Copyright (C) 2018-2019 Expedia, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.expediagroup.apiary.extensions.events.metastore.consumer.common.exception;
+
+/**
+ * An exception class representing runtime exceptions thrown by Hive Clients.
+ */
+public class HiveClientException extends RuntimeException {
+
+  public HiveClientException(String message, Throwable cause) {
+    super(message, cause);
+  }
+}
