+++ apiary-metastore-events/apiary-metastore-consumer/privileges-grantor/privileges-grantor-core/src/main/java/com/expediagroup/apiary/extensions/events/metastore/consumer/privilegesgrantor/core/Grantor.java
@@ -1,0 +1,31 @@
+/**
+ * Copyright (C) 2018-2019 Expedia, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.expediagroup.apiary.extensions.events.metastore.consumer.privilegesgrantor.core;
+
+public enum Grantor {
+  ADMIN("admin");
+
+  private final String grantor;
+
+  Grantor(String grantor) {
+    this.grantor = grantor;
+  }
+
+  @Override
+  public String toString() {
+    return grantor;
+  }
+}
