+++ apiary-metastore-events/apiary-metastore-consumer/privileges-grantor/privileges-grantor-lambda/src/main/java/com/expediagroup/apiary/extensions/events/metastore/consumer/privilegesgrantor/lambda/Response.java
@@ -1,0 +1,52 @@
+/**
+ * Copyright (C) 2018-2019 Expedia, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.expediagroup.apiary.extensions.events.metastore.consumer.privilegesgrantor.lambda;
+
+import com.google.gson.Gson;
+
+public class Response {
+
+  private int statusCode;
+  private String description;
+
+  public Response() {}
+
+  public Response(int statusCode, String description) {
+    this.statusCode = statusCode;
+    this.description = description;
+  }
+
+  public int getStatusCode() {
+    return statusCode;
+  }
+
+  public void setStatusCode(int statusCode) {
+    this.statusCode = statusCode;
+  }
+
+  public String getDescription() {
+    return description;
+  }
+
+  public void setDescription(String description) {
+    this.description = description;
+  }
+
+  public String toJsonString() {
+    Gson gson = new Gson();
+    return gson.toJson(this);
+  }
+}
