+++ pom.xml
@@ -32,6 +32,7 @@
     <mockito.version>2.21.0</mockito.version>
     <shade.prefix>${project.groupId}.shaded</shade.prefix>
     <slf4j.version>1.7.25</slf4j.version>
+    <beeju.version>1.3.0</beeju.version>
   </properties>
 
   <modules>
@@ -64,6 +65,12 @@
         <groupId>org.mockito</groupId>
         <artifactId>mockito-core</artifactId>
         <version>${mockito.version}</version>
+        <scope>test</scope>
+      </dependency>
+      <dependency>
+        <groupId>com.hotels</groupId>
+        <artifactId>beeju</artifactId>
+        <version>${beeju.version}</version>
         <scope>test</scope>
       </dependency>
     </dependencies>
